Code Cleanup
- Moved all remaining enums to the CssUI.Enums namespace.

- Finished multi-string property implementation.

- Altered all property classes to replace their Actual value with a typecasted version.

- All properties within the ElementPropertySystem that previously cached the final typecasted value for a given property have now been altered to simply serve as getters to the Actual value of their respective cascaded property.

+++ CssUI/Enums/ECssPropertyStage.cs
@@ -4,7 +4,7 @@ namespace CssUI
     /// <summary>
     /// Describes all of the different value stages for a Css property
     /// </summary>
-    public enum ECssPropertyStage : byte
+    public enum ECssPropertyStage : int
     {
         Invalid = 0,
         /// <summary>
